pyproject.toml
[tool.poetry]
name = "vetorizador"
version = "0.1.0"
description = ""
authors = ["nogueiramf <marcio.nogueira@rankmyapp.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-docx = "^1.1.2"
pymongo = "^4.8.0"
transformers = "^4.42.4"
torch = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
